fibonaci number

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -12,15 +12,15 @@
 int main(void)
 {
 	int x;
-	unsigned long long a = 0;
-	unsigned long long b = 1;
+	unsigned long a = 0;
+	unsigned long b = 1;
 
 	for (x = 1; x <= 50; x++)
 	{
 
-	unsigned long long fibo = a + b;
+	unsigned long fibo = a + b;
 
-		printf("%llu, ", fibo);
+		printf("%lu, ", fibo);
 
 		a = b;
 		b = fibo;
